- rename project folder - add sklearn pipeline with LDA and isolation forest
