+++ tx_class/scripts/lint.sh
