+++ tx_class/src/app/main.py
@@ -2,10 +2,10 @@ import time
 
 from ray import serve
 
-from fraud.app.conf import read_conf
-from fraud.app.model import ModelClassification
-from fraud.app.prediction import Prediction
-from fraud.app.service import app  # Import the FastAPI app
+from src.app.conf import read_conf
+from src.app.model import ModelClassification
+from src.app.prediction import Prediction
+from src.app.service import app  # Import the FastAPI app
 
 serve.start()
