+++ tx_class/tests/fixture_set.py
@@ -1,8 +1,6 @@
 import pytest
 from pyspark.sql import SparkSession
-
-from fraud.app.conf import read_conf
-from fraud.app.logger_custom import LoggerCustom
+from src.app.conf import read_conf
 
 
 @pytest.fixture(scope="module")
@@ -13,8 +11,3 @@ def conf():
 @pytest.fixture(scope="module")
 def spark():
     return SparkSession.builder.master("local[1]").appName("pytest").getOrCreate()
-
-#
-# @pytest.fixture(scope="module")
-# def logger():
-#     return LoggerCustom().logger
